Moved PublishRelease.java to the se.simonsoft.cms.publish.rest package
