+++ cms-publish-rest/src/main/java/se/simonsoft/cms/publish/rest/writers/PublishReleaseMessageBodyWriterHtml.java
@@ -25,7 +25,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import se.repos.web.ReposHtmlHelper;
 import se.simonsoft.cms.item.CmsItemId;
-import se.simonsoft.cms.publish.config.databinds.release.PublishRelease;
+import se.simonsoft.cms.publish.rest.PublishRelease;
 
 import javax.inject.Inject;
 import javax.ws.rs.Produces;
